third_party/blink/web_tests/media/mediasession/mojo/metadata-propagated-twice.html

<!DOCTYPE html>
<title>MediaSession Mojo Test</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/third_party/blink/public/mojom/mediasession/media_session.mojom.js"></script>
<script src="resources/mediasessionservice-mock.js"></script>
<script src="resources/utils.js"></script>
<script>

async_test(function(t) {
  let m = mediaSessionServiceMock;
  var dontCareMetadata = new MediaMetadata({});

  m.setMetadataCallback(t.step_func(function() {
    var metadata = new MediaMetadata({
      title: "title2",
      artist: "artist2",
      album: "album2",
      artwork: [
        { src: "http://foo.com/bar.jpg", type: "image/jpeg", sizes: "256x256"}
      ]});

    m.setMetadataCallback(t.step_func(function(receivedMetadata) {
      assert_metadata_equals(metadata, receivedMetadata);
      t.done();
    }));
    window.navigator.mediaSession.metadata = metadata;
  }));
  window.navigator.mediaSession.metadata = dontCareMetadata;
}, "test that MediaMetadata is correctly propagated twice");

</script>
